Renamed compiler.back to compiler.middle
