+++ src/main/java/org/manifold/compiler/InvalidIdentifierException.java
@@ -1,6 +1,6 @@
 package org.manifold.compiler;
 
-import org.manifold.compiler.back.SchematicException;
+import org.manifold.compiler.middle.SchematicException;
 
 public class InvalidIdentifierException extends SchematicException {
   private static final long serialVersionUID = -1571780250014515287L;
