+++ src/main/java/org/manifold/compiler/UndeclaredIdentifierException.java
@@ -1,6 +1,6 @@
 package org.manifold.compiler;
 
-import org.manifold.compiler.back.SchematicException;
+import org.manifold.compiler.middle.SchematicException;
 
 public class UndeclaredIdentifierException extends SchematicException {
   private static final long serialVersionUID = -5785755001929744865L;
