+++ src/test/java/org/manifold/compiler/TestNode.java
@@ -9,7 +9,7 @@ import org.junit.Test;
 
 import java.util.HashMap;
 import java.util.Map;
-import org.manifold.compiler.back.SchematicException;
+import org.manifold.compiler.middle.SchematicException;
 
 public class TestNode {
